Set default close operation to EXIT_ON_CLOSE
Changed the default close operation in InventoryView from DO_NOTHING_ON_CLOSE to JFrame.EXIT_ON_CLOSE to ensure the application exits when the window is closed. This was originally an artifact from HW 3 that was unintentionally carried over due to the import/export feature of the file-based storage system.